requirements.txt
streamlit==1.40.2
streamlit-authenticator==0.4.1
SQLAlchemy>=2.0.40,<2.1
openpyxl==3.1.5
bcrypt==4.2.1
PyYAML==6.0.2
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
